Split tmux pane row helpers
